# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Onyx Domain API V1",
    "version" : "1.0.9"
  },
  "servers" : [ {
    "url" : "http://onyxdomainapi.azurewebsites.net"
  } ],
  "paths" : {
...
    "/CashTransfers/List" : {
      "get" : {
        "tags" : [ "CashTransfers" ],
        "parameters" : [ {
          "name" : "f",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/CashTransfers/Download" : {
      "get" : {
        "tags" : [ "CashTransfers" ],
        "parameters" : [ {
          "name" : "f",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/octet-stream" : { }
            }
          }
        }
      }
    },
    "/CashTransfers/Detail" : {
      "get" : {
        "tags" : [ "CashTransfers" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/CashTransfers/Delete" : {
      "post" : {
        "tags" : [ "CashTransfers" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/CashTransfers/Save" : {
      "post" : {
        "tags" : [ "CashTransfers" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CashTransferData"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CashTransferData"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CashTransferData"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/ComissionsReport/List" : {
      "post" : {
        "tags" : [ "ComissionsReport" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ComissionParameters"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ComissionParameters"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ComissionParameters"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Context" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "nullable" : true
          },
          "success" : {
            "type" : "boolean"
          },
          "data" : {
            "nullable" : true
          }
        },
        "additionalProperties" : false,
        "example" : {
          "data" : "",
          "success" : true,
          "message" : "message"
        }
      },
...
      "IdOnly" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "additionalProperties" : false
      },
...
      "CashTransferData" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "date" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "settlementDate" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "buyValue" : {
            "type" : "number",
            "format" : "double"
          },
          "sellValue" : {
            "type" : "number",
            "format" : "double"
          },
          "buyAccountId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "sellAccountId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "fundId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
      "ComissionParameters" : {
        "type" : "object",
        "properties" : {
          "startDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "endDate" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "additionalProperties" : false
      },
...
    }
  }
}